rename open to _matterport_interactive.py
